Allow smooth compilation with MSVC++ 14.2 for the x64 arch
The MSVC compiler uses the LLP64 data model and, therefore, long is not
pointer-size wide like on Unix systems (using LP64). In this patch, `int_type`
is re-defined to `intptr_t` instead of `long`.

In addition to that, `size_type` is conditionally defined to `size_t` when
the project is built with MSVC, just to avoid tons of truncation warnings
or tons of explicit casts. Yes, that will make EvalValue bigger, but that's
fine. After all, this is a small educational project.

Finally, explicit casts have been used in a few places in order to avoid
some warnings and `Loc` uses now a template ctor in order to reduce the
need to write explicit casts to `int` in many places.

+++ src/mylang.cpp
@@ -60,7 +60,7 @@ read_script(const char *filename)
     }
 
     for (size_t i = 0; i < lines.size(); i++)
-        lexer(lines[i], i+1, tokens);
+        lexer(lines[i], static_cast<int>(i+1), tokens);
 }
 
 void
